sample_project/requirements.txt
-e ../
channels_redis